pyproject.toml
[project]
name = "cfpb-mcp"
version = "0.1.0"
description = "Hybrid FastAPI app that serves both MCP (/mcp) and REST endpoints for CFPB complaints."
requires-python = ">=3.10"

dependencies = [
    "fastapi>=0.125.0",
    "httpx>=0.28.1",
    "mcp>=1.24.0",
    "playwright>=1.48.0",
    "uvicorn[standard]>=0.38.0",
]

[tool.uv]

[dependency-groups]
dev = [
    "anthropic>=0.75.0",
    "anyio>=4.12.0",
    "jsonschema>=4.25.1",
    "openai>=2.13.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
]
# Keep the venv in-project as .venv
